SuperArchitect_Github/requirements.txt
requests
anthropic
openai
google-generativeai
aiohttp
asyncio
python-dotenv
PyYAML
faster-whisper
browsergym
click
Pillow